fix(report): move build_anchor_choices out of the PyQt6 widget module
CI installs no PyQt6 extra, so tests that imported build_anchor_choices from
widgets/annotation_dialog.py failed to collect there even though the function
itself needs no Qt. Move it to utils/anchors.py, a Qt-free module, and have
the dialog import it from there. Add a subprocess-based regression test that
imports the anchors module with PyQt6 blocked via sys.meta_path.
